fix(agent-pane): control-bar model/effort changes apply to running Claude
Verifying "setting the model (Sonnet/Opus) works" surfaced a gap: the
slash /model path got the #1503 persistent-restart fix, but the GUI
control-bar dropdown has its OWN updateRuntime that only wrote
`agent:runtime` meta — which silently no-ops for persistent (Claude
stream-json) controllers, since their --model/--effort are baked in at
spawn. So picking Opus/Sonnet in the dropdown didn't take effect on a
running Claude agent (launch-time + /model both worked).

- New runtime-apply.ts::applyRuntimeChange — the single apply path:
  write `agent:runtime`, and for persistent controllers rebuild `cmd:args`
  (buildRuntimeArgs) + force a ControllerResync (resumes via --resume,
  conversation preserved).
- commands/global/runtime.ts: /model·/effort·/mode now call the shared
  helper (de-duplicates the inline #1503 logic).
- AgentControlBar: updateRuntime calls applyRuntimeChange(getProvider(id))
  instead of a meta-only write — fixes the dropdown.

Verified: buildRuntimeArgs maps --model opus/sonnet correctly (10/10
tests); build clean.
