+++ .changesets/1781572000-fix-control-bar-model-apply.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+fix(agent-pane): model/effort/permission picked in the control-bar dropdown now applies to a running Claude agent — previously only the /model slash command did (the GUI write was meta-only and silently no-op'd for persistent controllers)
